Reorganize functions and fix search
