+++ chemistry/chemistry_system/urls.py
@@ -1,10 +1,9 @@
-from django.urls import path
 from chemistry_system import views
-from .views import login_view
 from chemistry_system.models import currentlyInStorageTable
 from django.contrib import admin
 from django.urls import path, include
 from django.contrib.auth import views as auth_views
+from chemistry_system.utils import export_chemicals_csv, import_chemicals_csv, update_checkout_status  # Import the new views
 
 curr_list_view = views.ChemListView.as_view(
     queryset=currentlyInStorageTable.objects.order_by("-chemBottleIDNUM"),
@@ -34,11 +33,11 @@ urlpatterns = [
     path('delete_chemical/<str:model_name>/<int:pk>/', views.delete_chemical, name='delete_chemical'),
     path('scanner_add/', views.scanner_add, name='scanner_add'),
     path('add/<str:model_name>/', views.add_chemical, name='add_chemical'),
-    path('export_chemicals_csv/', views.export_chemicals_csv, name='export_chemicals_csv'),
-    path('import_chemicals_csv/', views.import_chemicals_csv, name='import_chemicals_csv'),
-    path('update-checkout-status/<str:model_name>/<str:qrcode_value>/', views.update_checkout_status, name='update_checkout_status'),
+    path('export_chemicals_csv/', export_chemicals_csv, name='export_chemicals_csv'),  # Update the URL pattern
+    path('import_chemicals_csv/', import_chemicals_csv, name='import_chemicals_csv'),  # Update the URL pattern
+    path('update-checkout-status/<str:model_name>/<str:qrcode_value>/', update_checkout_status, name='update_checkout_status'),  # Update the URL pattern
     path('delete_all_chemicals/', views.delete_all_chemicals, name='delete_all_chemicals'),
     path('print/', views.print_page, name='print_page'),
-    path('download-qr-pdf/', views.generate_qr_pdf, name='download_qr_pdf'),
+    path('download-qr-pdf/', views.generate_qr_pdf_view, name='download_qr_pdf'),  # Update the URL pattern
     path('log/', views.log, name='log'),
 ]
